Moved files.  Fixed bug causing test failure
